Fix navigation to create a new realm and environment settings object
This fixes a regression introduced (apparently on purpose, without
understanding the implications) in cf0355d7e0e229b98f7fbd51b8c7608010c.
Now, realms and environment settings objects are properly 1:1 with
Window objects. See some discussion in #67 and in
https://github.com/whatwg/html/pull/322#issuecomment-197484831.

Drive-by fix: it's actually ParseScript that depends on the realm, not
ScriptEvaluation.