Clarify settings object, realm, and global relationships
See #167 and https://www.w3.org/Bugs/Public/show_bug.cgi?id=27204. This
makes it clear that environment settings objects, realms, and global
objects are 1:1:1, giving explicit names and paths for getting from each
to the other (of which all are already used in this spec). It also
introduces the "new" concept of "current settings object" as a shorthand
for "the current Realm Record's settings object", since that was already
being used in two places.

This does not yet fix #473, nor does it yet create definitions for
{incumbent, entry, current} {settings object, Realm, global object} like
I planned in https://github.com/whatwg/html/issues/167#issuecomment-195478361.